refactor: move IPC types to ipc-types.ts for better modularity
- Updated imports across multiple files to reference the new ipc-types.ts instead of ipc.ts.
- This change improves the separation of type definitions from the invocation logic, reducing unnecessary dependencies and potential circular imports.
- Adjusted related tests and logic files to ensure compatibility with the new structure.
