+++ apps/desktop/src/components/editor/properity-panel/captions-panel.logic.ts
@@ -4,7 +4,7 @@
  * picker's "active theme vs. Custom" readout.
  */
 
-import type { CaptionModelInfo, DeviceCapabilities } from "$lib/ipc";
+import type { CaptionModelInfo, DeviceCapabilities } from "$lib/ipc-types";
 import type { CaptionPresetValue } from "$lib/registry/types";
 import type { CaptionStyle } from "$lib/stores/editor-store.svelte";
 import { resolveCaptionAnimation } from "@recast/captions";
