+++ apps/desktop/src/components/editor/properity-panel/dev-ocr-panel.logic.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import type { OcrProgress, OcrStats, ScreenStateSpan, VideoTextTimeline } from "$lib/ipc";
+import type { OcrProgress, OcrStats, ScreenStateSpan, VideoTextTimeline } from "$lib/ipc-types";
 import {
 	boxLabel,
 	boxRect,
