+++ packages/media/test/perf/bundle.test.ts
@@ -1,19 +1,22 @@
 import { build } from 'esbuild';
+import { dirname, resolve } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { gzipSync } from 'node:zlib';
 import { describe, expect, it } from 'vitest';
 
 /**
- * Bundle-size gates for REQUIREMENTS.md §3. These are the latency-free budget
- * rows, so unlike the rest they can be enforced without a browser.
+ * Bundle-size gates for REQUIREMENTS.md §3 — the only budget rows checkable
+ * without a browser. Sizes are gzipped KB of a minified, tree-shaken bundle.
  */
 
+const PKG_ROOT = resolve(dirname(fileURLToPath(import.meta.url)), '../..');
 const DESKTOP_BUDGET_GZ_KB = 80;
-const WEB_BUDGET_GZ_KB = 150;
+const WORKER_BUDGET_GZ_KB = 220;
 
 /** Bundle `code` as a virtual entry and return its gzipped size in KB. */
 async function bundleGzKb(code: string): Promise<number> {
 	const result = await build({
-		stdin: { contents: code, resolveDir: new URL('../../', import.meta.url).pathname, loader: 'ts' },
+		stdin: { contents: code, resolveDir: PKG_ROOT, loader: 'ts' },
 		bundle: true,
 		format: 'esm',
 		platform: 'browser',
@@ -29,34 +32,41 @@ async function bundleGzKb(code: string): Promise<number> {
 }
 
 describe('bundle budgets (REQUIREMENTS.md §3)', () => {
-	it('the desktop playback surface stays under budget', async () => {
+	it('the desktop surface stays under budget', async () => {
 		const kb = await bundleGzKb(`
 			import { getFrameCache, MediaError, isUnsupportedContainer } from './src/index';
-			console.log(getFrameCache, MediaError, isUnsupportedContainer);
+			import { MediabunnyVideoSource } from './src/playback/index';
+			console.log(getFrameCache, MediaError, isUnsupportedContainer, MediabunnyVideoSource);
 		`);
 		expect(kb).toBeLessThanOrEqual(DESKTOP_BUDGET_GZ_KB);
 	});
 
-	it('the web conversion surface stays under budget', async () => {
+	it('the barrel is tree-shakable', async () => {
+		// Regression guard, two ways: MediaBunny used to be re-exported from the
+		// barrel, and the package never declared `sideEffects: false`. Either
+		// alone made a lone `MediaError` import cost 61 KB gz instead of 0.2.
 		const kb = await bundleGzKb(`
-			import { runConversion, outputFormatFor, handlers } from './src/index';
-			console.log(runConversion, outputFormatFor, handlers);
-		`);
-		expect(kb).toBeLessThanOrEqual(WEB_BUDGET_GZ_KB);
-	});
-
-	it('the main barrel does not drag MediaBunny into every consumer', async () => {
-		// Regression: MediaBunny used to be re-exported from `src/index.ts`, so a
-		// consumer importing only `MediaError` still paid for the whole library.
-		const errorOnly = await bundleGzKb(`
 			import { MediaError } from './src/index';
 			console.log(MediaError);
 		`);
-		const withMediabunny = await bundleGzKb(`
-			import { Input } from './src/mediabunny';
-			console.log(Input);
+		expect(kb).toBeLessThan(5);
+	});
+
+	it('the playback subpath does not pull the conversion pipeline', async () => {
+		const kb = await bundleGzKb(`
+			import { MediabunnyVideoSource } from './src/playback/index';
+			console.log(MediabunnyVideoSource);
 		`);
-		expect(errorOnly).toBeLessThan(withMediabunny / 2);
-		expect(errorOnly).toBeLessThan(20);
+		expect(kb).toBeLessThan(20);
+	});
+
+	it('the conversion worker surface stays under budget', async () => {
+		// On-demand chunk: apps/web's client.ts is types-only and spawns the
+		// worker lazily, so this never blocks first paint.
+		const kb = await bundleGzKb(`
+			import { handlers, runConversion, outputFormatFor } from './src/index';
+			console.log(handlers, runConversion, outputFormatFor);
+		`);
+		expect(kb).toBeLessThanOrEqual(WORKER_BUDGET_GZ_KB);
 	});
 });
